+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ENHANCEMENTS:
 * Upgrade to [v2 of the Terraform Plugin SDK Go module](https://developer.hashicorp.com/terraform/plugin/sdkv2/guides/v2-upgrade-guide#version-2-of-the-module)
 
+BUG FIXES:
+* Fix a typo when reading a sumologic_cse_entity_normalization_configuration
+
 ## 3.0.12 (May 21, 2025)
 
 ## 3.0.11 (May 20, 2025)